Fix problem with type

+++ packages/parser/src/sections/error.rs
@@ -4,5 +4,5 @@ use serde::{Deserialize, Serialize};
 pub struct ERROR {
     pub message: String,
     pub line: String,
-    pub line_number: usize,
+    pub line_number: u32,
 }
